Add modified code and project description
n this question you will write several R functions for working with “mouse-tracking” data as described in this manuscript. Briefly, suppose you have data in the form of a series of triples (x, y, t) representing the position (x, y) of a mouse cursor in the plane (i.e. monitor) at time t during a trial in which participants click one of two buttons in response to a prompt. The buttons are arranged horizontally on opposite sides of the screen to allow the experimenter to garner information about how decisively each response is given.

For each part, write an R function to accomplish the stated task. Name each function using an informative verb. Use comments to document the arguments and output of each function. Be sure to clearly state any assumptions on the inputs.

Submit your answers to parts a-e as a single executable R script ps1_q3.R. Also submit a pdf created using Rmarkdown with answers to all parts including your function definitions.

[5pts] Write a function that accepts a n n×3 matrix representing the trajectory (x, y, t) and translates it to begin with time zero at the origin.

[5pts] Write a function that computes the angle θ formed by the secant line connecting the origin and the final position in the trajectory. Your answer should be an angle between [−π,π]. Be sure your your solution works for a trajectory ending in any of the four quadrants.

[5pts] Write a function to rotate the (x, y) coordinates of a trajectory so that the final point lies along the positive x-axis.

[2pts] Combine the three parts above into a single function that normalizes an n×3 trajectory matrix to begin at the origin and end on the positive x-axis.

[8pts] Write a function that accepts a normalized trajectory and computes the following metrics describing its curvature:
the total (Euclidean) distance traveled,
the maximum absolute deviation from the secant connecting the starting and final positions,
the average absolute deviation of the observed trajectory from the direct path,
the (absolute) area under the curve for the trajectory relative to the secant line using the trapezoidal rule to integrate. [Hint: Allow cancellation in “x” but not “y”.]
[5pts] Apply your function to the sample trajectories at the Stats506_F19 repo on GitHub and check your solutions against the sample measures. Then, compute the metrics above for the test trajectories and report your results in a nicely formatted table.